Fix broken imports after component restructure
- Update legend component to import from primitives/private
- Update legend stories to import charts from primitives
- Update leaderboard-chart hook to import from components/legend
- Update sample-data to import types from primitives
- Fix storybook title in legend stories
