+++ projects/js-packages/charts/src/components/legend/legend.tsx
@@ -1,6 +1,6 @@
 import { useContext, useMemo, forwardRef } from 'react';
+import { SingleChartContext } from '../../primitives/private/single-chart-context';
 import { GlobalChartsContext } from '../../providers';
-import { SingleChartContext } from '../private/single-chart-context';
 import { BaseLegend } from './private';
 import type { LegendProps } from './types';
